move demos to examples directory outside of packages
