+++ examples/browser-embedded-sdk-demo-app/CHANGELOG.md
