+++ examples/browser-embedded-sdk-demo-app/src/App.tsx
