+++ examples/browser-embedded-sdk-demo-app/src/ErrorBoundary.tsx
