+++ examples/browser-embedded-sdk-demo-app/src/components/ElementConfigScreen.tsx
@@ -1,36 +1,40 @@
 import * as React from "react";
-import type { Phantom } from "../../../../browser-embedded-sdk/src/index";
+import type { Phantom } from "../../../../packages/browser-embedded-sdk/src/index";
+import { Prism, type SyntaxHighlighterProps } from "react-syntax-highlighter";
+
 import { docco } from "react-syntax-highlighter/dist/esm/styles/hljs";
 import { WalletControls } from "./WalletControls";
-import { Prism, type SyntaxHighlighterProps } from "react-syntax-highlighter";
 const SyntaxHighlighter = Prism as any as React.FC<SyntaxHighlighterProps>;
 
-interface NormalConfigScreenProps {
+interface ElementConfigScreenProps {
   phantom: Phantom | null;
   onBack: () => void;
+  children: React.ReactNode;
 }
 
-export const NormalConfigScreen: React.FC<NormalConfigScreenProps> = ({ phantom, onBack }) => {
+export const ElementConfigScreen: React.FC<ElementConfigScreenProps> = ({ phantom, onBack, children }) => {
   const codeSnippet = `
-import { createPhantom, Position } from "@phantom/wallet-sdk";
+import { createPhantom } from "@phantom/wallet-sdk";
 
-// Normal configuration (bottom right)
+// Element-based configuration
 const phantom = await createPhantom({
-  position: Position.bottomRight,
-  hideLauncherBeforeOnboarded: false,
-  namespace: "normal-wallet",
+  element: "wallet-container", // ID of the container element
+  namespace: "element-wallet",
 });
   `;
 
   return (
-    <div className="config-screen">
-      <h2>Normal Wallet Configuration</h2>
+    <div className="config-screen" id="element-wallet-page">
+      <h2>Element-Based Wallet Configuration</h2>
       <div className="code-snippet">
         <h3>Integration Code</h3>
         <SyntaxHighlighter language="typescript" style={docco}>
           {codeSnippet}
         </SyntaxHighlighter>
       </div>
+
+      <p>The wallet will render inside the container below:</p>
+      {children}
 
       <WalletControls phantom={phantom} />
