+++ examples/browser-embedded-sdk-demo-app/src/components/NavigationScreen.tsx
