+++ examples/browser-embedded-sdk-demo-app/src/components/WalletContainer.tsx
