+++ examples/browser-embedded-sdk-demo-app/src/index.css
