+++ examples/browser-embedded-sdk-demo-app/src/main.tsx
