+++ examples/browser-embedded-sdk-demo-app/tsconfig.node.json
@@ -1,5 +1,5 @@
 {
-  "extends": "../../../tsconfig.json",
+  "extends": "../../tsconfig.json",
   "compilerOptions": {
     "composite": true,
     "skipLibCheck": true,
