+++ examples/browser-sdk-demo-app/index.html
