+++ examples/browser-sdk-demo-app/package.json
@@ -4,9 +4,9 @@
   "version": "0.0.3",
   "type": "module",
   "scripts": {
-    "predev": "cd ../../browser-sdk && yarn build:watch",
+    "predev": "yarn workspace @phantom/browser-sdk build:watch",
     "dev": "concurrently \"yarn predev\" \"vite --port 5175\"",
-    "prebuild": "cd ../../browser-sdk && yarn build",
+    "prebuild": "yarn workspace @phantom/browser-sdk build",
     "build": "tsc && vite build",
     "preview": "vite preview"
   },
