+++ examples/browser-sdk-demo-app/src/main.ts
