+++ examples/browser-sdk-demo-app/src/style.css
