+++ examples/browser-sdk-demo-app/tsconfig.json
