+++ examples/browser-sdk-demo-app/vite.config.ts
