+++ examples/react-sdk-demo-app/src/vite-env.d.ts
