+++ examples/react-sdk-demo-app/vite.config.ts
